requirements.txt
google-generativeai
groq
requests
beautifulsoup4
PyPDF2
python-dotenv
termcolor
torch
transformers
accelerate
hf_transfer
anthropic
